+++ src/main/java/insane96mcp/experienceoverhaul/network/NetworkHandler.java
@@ -1,7 +1,6 @@
 package insane96mcp.experienceoverhaul.network;
 
 import insane96mcp.experienceoverhaul.network.message.AnvilRepairSyncMessage;
-import insane96mcp.experienceoverhaul.network.message.SyncExperienceDisabledGameruleMessage;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
 import net.neoforged.neoforge.network.registration.PayloadRegistrar;
@@ -10,7 +9,6 @@ public class NetworkHandler {
     @SubscribeEvent
     public static void register(final RegisterPayloadHandlersEvent event) {
         final PayloadRegistrar registrar = event.registrar("1");
-        registrar.playToClient(SyncExperienceDisabledGameruleMessage.TYPE, SyncExperienceDisabledGameruleMessage.STREAM_CODEC, SyncExperienceDisabledGameruleMessage::handle);
         registrar.playToClient(AnvilRepairSyncMessage.TYPE, AnvilRepairSyncMessage.STREAM_CODEC, AnvilRepairSyncMessage::handle);
     }
 }