+++ changelog.md
@@ -1,3 +1,6 @@
+## Upcoming
+* Removed the game rule to disable experience
+
 ## 1.1.0.0-alpha
 * Ported Anvil features
   * Repairing items no longer costs more materials if the item is enchanted, but will instead cost 0.5 levels per enchantment level on it